Ajout de addCommentinMedia & deleteCommentinMedia

+++ src/components/MediaUpload.jsx
@@ -1,13 +1,10 @@
 // A MODIFIER AVEC BON CSS
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import { Link, useNavigate } from "react-router-dom";
+import { Link } from "react-router-dom";
 import Card from "./Card";
 
 function MediaUpload() {
-  const userData = JSON.parse(sessionStorage.getItem("user"));
-  const decoData = JSON.parse(sessionStorage.getItem("decorator"));
-
   const [selectedImage, setSelectedImage] = useState([]);
   const [selectedMediaId, setSelectedMediaId] = useState(null);
   const [selectedFilter, setSelectedFilter] = useState(null);
@@ -26,6 +23,7 @@ function MediaUpload() {
   const [modal, setModal] = useState(false);
   const [showConfirmationModal, setShowConfirmationModal] = useState(false);
 
+  const userData = JSON.parse(sessionStorage.getItem("user"));
   const decoratorData = JSON.parse(sessionStorage.getItem("decorator"));
 
   useEffect(() => {
@@ -43,7 +41,10 @@ function MediaUpload() {
           createAt: new Date(media.createAt),
         }))
         .sort((a, b) => b.createAt - a.createAt);
+
       setAllMedia(mediaData);
+      setCommentaire(mediaData);
+      console.log(mediaData);
       setOriginalMedia(mediaData); // Définir originalMedia après avoir défini allMedia
     } catch (error) {
       console.error("Error fetching media:", error);
@@ -77,10 +78,22 @@ function MediaUpload() {
     formData.append("viewUrl", viewUrl);
     console.log(selectedImage, title, description, viewUrl);
     try {
+      const token = JSON.parse(sessionStorage.getItem("token"));
+      console.log(token);
+      if (!token) {
+        throw new Error("Token not found");
+      }
+      const config = {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      };
       const response = await axios.post(
         "http://localhost:6789/media/create-media",
         formData,
-        { headers: { "Content-Type": "multipart/form-data" } }
+        {
+          headers: { ...config.headers, "Content-Type": "multipart/form-data" },
+        }
       );
 
       console.log("response", response.data);
@@ -134,7 +147,7 @@ function MediaUpload() {
     e.preventDefault();
     try {
       const token = JSON.parse(sessionStorage.getItem("token"));
-      console.log(token)
+      console.log(token);
       if (!token) {
         throw new Error("Token not found");
       }
@@ -158,8 +171,19 @@ function MediaUpload() {
   //---DELETE-----//
   const deleteMedia = async (mediaId) => {
     try {
+      const token = JSON.parse(sessionStorage.getItem("token"));
+      console.log(token);
+      if (!token) {
+        throw new Error("Token not found");
+      }
+      const config = {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      };
       const response = await axios.delete(
-        `http://localhost:6789/media/delete/${mediaId}`
+        `http://localhost:6789/media/delete/${mediaId}`,
+        config
       );
       console.log(response.data);
       getAllMedia();
@@ -196,27 +220,39 @@ function MediaUpload() {
     }
   };
 
-  const handleCommentaire = async (media, newCommentaire, author) => {
-    console.log(decoData.user.firstname);
-    console.log(userData.user.firstname);
+  const handleCommentaire = async (media, newCommentaire, author, createAt) => {
     try {
+      const token = JSON.parse(sessionStorage.getItem("token"));
+      if (!token) {
+        throw new Error("Token not found");
+      }
+      const config = {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      };
       const response = await axios.post(
         `http://localhost:6789/media/${media}/comment`,
         {
           commentaire: newCommentaire,
           author: author,
-        }
+          createAt,
+        },
+        config
       );
-      console.log("Comment added:", response.data);
+
+      console.log("Commentaire ajouté à :", response.data);
       getAllMedia();
       setCommentaire("");
-    } catch (error) {}
+    } catch (error) {
+      // Gérer les erreurs en conséquence
+    }
   };
 
-  const deleteComment = async (mediaId, comment) => {
+  const deleteComment = async (media, comment) => {
     try {
       const response = await axios.delete(
-        `http://localhost:6789/${mediaId}/delete-comment/${comment}`
+        `http://localhost:6789/media/${media}/delete-comment/${comment}`
       );
       console.log(response.data);
       getAllMedia();
@@ -261,7 +297,7 @@ function MediaUpload() {
             <input
               className="input-field-upload"
               type="text"
-              placeholder="Description"
+              placeholder="Titre"
               value={description}
               onChange={(e) => setDescription(e.target.value)}
             />
@@ -353,9 +389,26 @@ function MediaUpload() {
                   {media.commentaire.map((comment) => (
                     <div key={comment._id}>
                       <div className="box-one-comment">
-                        <p>{comment.author}</p>
-                        <p>{comment.commentaire}</p>
-                        <button onClick={() => deleteComment(comment._id)}>
+                        <p
+                          className="name-comment"
+                          style={{ fontSize: "10px", fontStyle: "italic" }}
+                        >
+                          {comment.author} :{" "}
+                        </p>
+                        <p
+                          className="content-comment"
+                          style={{ fontSize: "9px", fontStyle: "italic" }}
+                        >
+                          {comment.commentaire}
+                        </p> 
+                        <br />
+                        <br />
+                        <br />
+                        <p style={{ fontSize: "9px", fontStyle: "italic" }}>Crér à :{comment.createAt}</p>
+                        <button
+                          style={{ fontSize: "9px" }}
+                          onClick={() => deleteComment(media._id, comment._id)}
+                        >
                           x
                         </button>
                       </div>
@@ -470,233 +523,3 @@ function MediaUpload() {
 }
 
 export default MediaUpload;
-
-// // import React, { useState, useEffect } from "react";
-// // import axios from "axios";
-// // import { Link, useNavigate } from "react-router-dom";
-
-// // function MediaUpload() {
-// //   const [selectedImage, setSelectedImage] = useState([]);
-// //   const [title, setTitle] = useState("");
-// //   const [description, setDescription] = useState("");
-// //   const [viewUrl, setViewUrl] = useState("");
-// //   const [allImage, setAllImage] = useState(null);
-// //   const navigate = useNavigate();
-// //   const decoratorData = JSON.parse(sessionStorage.getItem("decorator"));
-
-// //   useEffect(() => {
-// //     getAllImage();
-// //   }, []);
-
-// //   const getAllImage = async () => {
-// //     const result = await axios.get("http://localhost:6789/media/allmedia");
-// //     console.log(result.data);
-// //     setAllImage(result.data);
-// //   };
-
-// //   const handleImageChange = (e) => {
-// //     console.log(e.target.files);
-// //     setSelectedImage(e.target.files);
-// //   };
-
-// //   const handleView = (e) => {
-// //     setViewUrl();
-// //     window.location.reload();
-// //   };
-
-// //   const handleUpload = async (e) => {
-// //     e.preventDefault();
-// //     const formData = new FormData();
-// //     for (const file of selectedImage) {
-// //       formData.append("mediaType", file);
-// //     }
-// //     formData.append("title", title);
-// //     formData.append("description", description);
-// //     formData.append("viewUrl", viewUrl);
-// //     console.log(selectedImage, title, description, viewUrl);
-// //     try {
-// //       const response = await axios.post(
-// //         "http://localhost:6789/media/create-media",
-// //         formData,
-// //         { headers: { "Content-Type": "multipart/form-data" } }
-// //       );
-// //       console.log("response", response.data);
-// //       if (response.data && response.data.media) {
-// //         const mediaData = response.data.media;
-// //         console.log("Média créé avec succès :", mediaData);
-// //       } else {
-// //         console.error("La réponse du serveur est incorrecte ou vide");
-// //       }
-// //       window.location.reload()
-// //       navigate("/decorators/espace-creation");
-// //       getAllImage();
-// //     } catch (error) {
-// //       console.error("Error uploading image:", error);
-// //     }
-// //   };
-
-// //   const deleteMedia = async (mediaId) => {
-// //     const result = await axios.delete(
-// //       `http://localhost:6789/media/delete/${mediaId}`
-// //     );
-// //     getAllImage();
-// //   };
-
-// //   const updateMedia = async (mediaId, newData) => {
-// //     try {
-// //       const response = await axios.put(
-// //         `http://localhost:6789/media/update-media/${mediaId}`,
-// //         newData
-// //       );
-// //       console.log("Updated media:", response.data);
-// //       getAllImage();
-// //     } catch (error) {
-// //       console.error("Error updating media:", error);
-// //     }
-// //   };
-
-// //   const handleUpdateTitle = (mediaId, newTitle) => {
-// //     updateMedia(mediaId, { title: newTitle });
-// //   };
-
-// //   const handleUpdateDescription = (mediaId, newDescription) => {
-// //     updateMedia(mediaId, { description: newDescription });
-// //   };
-
-// //   const handleUpdateView = async (mediaId, newView) => {
-// //     updateMedia(mediaId, { viewUrl: newView });
-// //     setViewUrl(newView);
-// //     window.location.reload();
-// //   };
-// //   return (
-// //     <>
-// //       <div className="container-form">
-// //         {decoratorData ? (
-// //           <form
-// //             enctype="multipart/form-data"
-// //             onSubmit={handleUpload}
-// //             className="form-user"
-// //           >
-// //             <input
-// //               className="input-field"
-// //               type="text"
-// //               placeholder="Titre"
-// //               value={title}
-// //               onChange={(e) => setTitle(e.target.value)}
-// //             />
-// //             <input
-// //               className="input-field"
-// //               type="text"
-// //               placeholder="Description"
-// //               value={description}
-// //               onChange={(e) => setDescription(e.target.value)}
-// //             />
-// //             <input
-// //               className="input-field"
-// //               type="text"
-// //               placeholder="Lien 360°"
-// //               value={viewUrl}
-// //               onChange={(e) => setViewUrl(e.target.value)}
-// //             />
-// //             <input type="file" name="mediaType" onChange={handleImageChange} />
-// //             <button type="submit">Poster fichier</button>
-// //           </form>
-// //         ) : null}
-// //       </div>
-// //       <h1>VOTRE PROJET</h1>
-// //       <div className="container-grid">
-// //         {allImage == null
-// //           ? ""
-// //           : allImage.map((image) => {
-// //               return (
-// //                 <div key={image._id}>
-// //                   <div className="card">
-// //                     <Link to={`/decorators/espace-creation/${image._id}`}>
-// //                       <img
-// //                         className="img-card"
-// //                         src={`http://localhost:6789/${image.mediaType}`}
-// //                       />
-// //                     </Link>
-// //                     <div className="box-infos">
-// //                       <h4 className="card-title">{image.title}</h4>
-
-// //                       <div className="box-second">
-// //                         <p className="card-description">{image.description}</p>
-
-// //                         <button
-// //                           className="btn-view"
-// //                           onClick={() => {
-// //                             window.open(image.viewUrl, viewUrl, "_blank");
-// //                           }}
-// //                         >
-// //                           View 360°
-// //                         </button>
-// //                       </div>
-
-// //                       {decoratorData ? (
-// //                         <div className="box-update">
-// //                           <div className="update-title">
-// //                             <input
-// //                               type="text"
-// //                               placeholder="Nouveau Titre"
-// //                               onChange={(e) => setTitle(e.target.value)}
-// //                             />
-// //                             <button
-// //                               onClick={() =>
-// //                                 handleUpdateTitle(image._id, title)
-// //                               }
-// //                             >
-// //                               Update
-// //                             </button>
-// //                           </div>
-// //                           <div className="update-description">
-// //                             <input
-// //                               type="text"
-// //                               placeholder="Nouvelle description"
-// //                               onChange={(e) => setDescription(e.target.value)}
-// //                             />
-// //                             <button
-// //                               onClick={() =>
-// //                                 handleUpdateDescription(image._id, description)
-// //                               }
-// //                             >
-// //                               Update
-// //                             </button>
-// //                           </div>
-
-// //                           <div className="update-description">
-// //                             <div>
-// //                               <input
-// //                                 type="text"
-// //                                 placeholder="Lien Url 360°"
-// //                                 onChange={(e) => setViewUrl(e.target.value)}
-// //                               />
-// //                               <button
-// //                                 onClick={() =>
-// //                                   handleUpdateView(image._id, viewUrl)
-// //                                 }
-// //                               >
-// //                                 Update View
-// //                               </button>
-// //                             </div>
-
-// //                             <button
-// //                               className="btn-delete"
-// //                               onClick={() => deleteMedia(image._id)}
-// //                             >
-// //                               Delete
-// //                             </button>
-// //                           </div>
-// //                         </div>
-// //                       ) : null}
-// //                     </div>
-// //                   </div>
-// //                 </div>
-// //               );
-// //             })}
-// //       </div>
-// //     </>
-// //   );
-// // }
-
-// // export default MediaUpload;
